requirements.txt
Pillow==10.2.0
mysql-connector-python==8.3.0
matplotlib==3.7.1 